Simplify our use of Huey in tests
Instead of re-implementing the Huey `execute()` method, we can call it
and use Signals to catch and re-raise exceptions.  This means we're
testing more of our use of Huey, and is overall simpler.  While here,
let's replace the complicated `tasks.run_pipeline_stages(N)` method with
a simpler `tasks.run_queued_tasks_and_enqueue_dependents()`.

The only downside is that the tracebacks of exceptions raised in our
Tasks aren't as easy to follow, but they still seem to contain all of
the important bits.